+++ exporters/variables-scss/tests/fixtures/mockedExampleTokens.ts
@@ -1,5 +1,4 @@
 import {
-  ColorToken,
   DimensionToken,
   DimensionTokenValue,
   StringToken,
@@ -7,7 +6,6 @@ import {
   Token,
   TokenGroup,
   TokenType,
-  TypographyToken,
   Unit,
 } from '@supernovaio/sdk-exporters';
 import { TokenGroupRemoteModel } from '@supernovaio/sdk-exporters/build/sdk-typescript/src/model/groups/SDKTokenGroup';
@@ -45,47 +43,7 @@ exampleMockedTokens.set('stringRef', {
   value: testString,
 } as StringToken);
 
-export const exampleMockedTypographyTokens = new Map<string, Token>();
-exampleMockedTypographyTokens.set('typographyRef1', {
-  id: 'typographyRef1',
-  name: 'Bold',
-  tokenType: TokenType.typography,
-  parentGroupId: '1',
-  origin: {
-    name: 'Heading/Desktop/XLarge/Bold',
-  },
-  value: {
-    fontSize: { unit: 'Pixels', measure: 64 },
-    lineHeight: { measure: 120 },
-    fontFamily: { text: 'Inter' },
-    fontWeight: { text: '700' },
-  },
-} as TypographyToken);
-exampleMockedTypographyTokens.set('typographyRef2', {
-  id: 'typographyRef2',
-  name: 'Bold-Underline',
-  tokenType: TokenType.typography,
-  parentGroupId: '2',
-  origin: {
-    name: 'Heading/Desktop/XLarge/Bold-Underline',
-  },
-  value: {
-    fontSize: { unit: 'Pixels', measure: 64 },
-    lineHeight: { measure: 120 },
-    fontFamily: { text: 'Inter' },
-    fontWeight: { text: '700' },
-  },
-} as TypographyToken);
-
-export const expectedTypographyValue = `(
-font-family: "'Inter', sans-serif",
-font-size: 64px,
-font-style: italic,
-font-weight: 700,
-line-height: 1.2
-)`;
-
-const groupFunctions = {
+export const groupFunctions = {
   addChild(): void {},
   addChildren(): void {},
   setParentGroupId(): void {},
@@ -178,136 +136,6 @@ export const exampleMockedGroups: TokenGroup[] = [
     tokenIds: ['stringRef'],
     subgroupIds: [],
     parentGroupId: 'parent3',
-    sortOrder: -1,
-    createdAt: null,
-    updatedAt: null,
-  },
-];
-
-export const exampleMockedColorsTokens = new Map<string, Token>();
-exampleMockedColorsTokens.set('actionColorRef', {
-  id: 'actionColorRef',
-  name: 'active',
-  tokenType: TokenType.color,
-  parentGroupId: '1',
-  origin: {
-    name: 'action/button/primary/default',
-  },
-  value: {
-    color: {
-      r: 202,
-      g: 32,
-      b: 38,
-      referencedTokenId: null,
-    },
-    opacity: {
-      unit: 'Raw',
-      measure: 1,
-      referencedTokenId: null,
-    },
-    referencedTokenId: null,
-  },
-} as ColorToken);
-exampleMockedColorsTokens.set('backgroundColorRef', {
-  id: 'backgroundColorRef',
-  name: 'primary',
-  tokenType: TokenType.color,
-  parentGroupId: '2',
-  origin: {
-    name: 'background/primary',
-  },
-  value: {
-    color: {
-      r: 255,
-      g: 255,
-      b: 255,
-      referencedTokenId: null,
-    },
-    opacity: {
-      unit: 'Raw',
-      measure: 1,
-      referencedTokenId: null,
-    },
-    referencedTokenId: null,
-  },
-} as ColorToken);
-
-export const exampleMockedColorGroups: TokenGroup[] = [
-  {
-    ...groupFunctions,
-    id: '1',
-    idInVersion: 'idInVersionValue',
-    brandId: 'brandIdValue',
-    designSystemVersionId: 'designSystemVersionIdValue',
-    name: 'primary',
-    description: '',
-    isRoot: false,
-    tokenType: TokenType.color,
-    childrenIds: ['actionColorRef'],
-    path: ['action', 'button'],
-    tokenIds: ['actionColorRef'],
-    subgroupIds: [],
-    parentGroupId: 'parent1',
-    sortOrder: -1,
-    createdAt: null,
-    updatedAt: null,
-  },
-  {
-    ...groupFunctions,
-    id: '2',
-    idInVersion: 'idInVersionValue',
-    brandId: 'brandIdValue',
-    designSystemVersionId: 'designSystemVersionIdValue',
-    name: 'background',
-    description: '',
-    isRoot: false,
-    tokenType: TokenType.color,
-    childrenIds: ['backgroundColorRef'],
-    path: [],
-    tokenIds: ['backgroundColorRef'],
-    subgroupIds: [],
-    parentGroupId: 'parent2',
-    sortOrder: -1,
-    createdAt: null,
-    updatedAt: null,
-  },
-];
-
-export const exampleMockedTypographyGroups: TokenGroup[] = [
-  {
-    ...groupFunctions,
-    id: '1',
-    idInVersion: 'idInVersionValue',
-    brandId: 'brandIdValue',
-    designSystemVersionId: 'designSystemVersionIdValue',
-    name: 'Heading',
-    description: '',
-    isRoot: false,
-    tokenType: TokenType.typography,
-    childrenIds: ['typographyRef1'],
-    path: [],
-    tokenIds: ['typographyRef1'],
-    subgroupIds: [],
-    parentGroupId: 'parent1',
-    sortOrder: -1,
-    createdAt: null,
-    updatedAt: null,
-  },
-  {
-    ...groupFunctions,
-    id: '2',
-    idInVersion: 'idInVersionValue',
-    brandId: 'brandIdValue',
-    designSystemVersionId: 'designSystemVersionIdValue',
-    name: 'Heading',
-    description: '',
-    isRoot: false,
-    tokenType: TokenType.typography,
-    childrenIds: ['typographyRef2'],
-    path: [],
-    tokenIds: ['typographyRef2'],
-    subgroupIds: [],
-    parentGroupId: 'parent2',
     sortOrder: -1,
     createdAt: null,
     updatedAt: null,
